Revamp of binary tree- deletes class node, creates staic class
