+++ .idea/workspace.xml
@@ -2,11 +2,13 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="e600a3aa-b72b-4e12-b49a-77194d49168f" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/src/BinaryTree/Tree.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/BinaryTree/TreeTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/Data_Structure_Practice/BinaryTree/Node.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/Data_Structure_Practice/BinaryTree/Node.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/Data_Structure_Practice/BinaryTree/Node.class" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/out/production/Data_Structure_Practice/BinaryTree/NodeTest.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/Data_Structure_Practice/BinaryTree/NodeTest.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/BinaryTree/Node.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/BinaryTree/Node.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/BinaryTree/NodeTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/BinaryTree/NodeTest.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/BinaryTree/Node.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/BinaryTree/NodeTest.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -53,22 +55,7 @@
       <recent name="" />
     </key>
   </component>
-  <component name="RunManager" selected="JUnit.NodeTest.canNodesBeTraversed">
-    <configuration name="NodeTest" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
-      <module name="Data_Structure_Practice" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="BinaryTree.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <option name="PACKAGE_NAME" value="BinaryTree" />
-      <option name="MAIN_CLASS_NAME" value="BinaryTree.NodeTest" />
-      <option name="TEST_OBJECT" value="class" />
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
+  <component name="RunManager" selected="JUnit.TreeTest.doesNodeAndValueExist">
     <configuration name="NodeTest.canNodesBeTraversed" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
       <module name="Data_Structure_Practice" />
       <extension name="coverage">
@@ -133,13 +120,29 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
+    <configuration name="TreeTest.doesNodeAndValueExist" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="Data_Structure_Practice" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="BinaryTree.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="BinaryTree" />
+      <option name="MAIN_CLASS_NAME" value="BinaryTree.TreeTest" />
+      <option name="METHOD_NAME" value="doesNodeAndValueExist" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="JUnit.TreeTest.doesNodeAndValueExist" />
         <item itemvalue="JUnit.NodeTest.canNodesBeTraversed" />
         <item itemvalue="JUnit.NodeTest.canYouAddNodes" />
         <item itemvalue="JUnit.NodeTest.doesNodeAndValueExist" />
         <item itemvalue="JUnit.NodeTest.doesNodeExist" />
-        <item itemvalue="JUnit.NodeTest" />
       </list>
     </recent_temporary>
   </component>
@@ -165,29 +168,29 @@
     <option name="version" value="1" />
   </component>
   <component name="WindowStateProjectService">
-    <state width="1398" height="451" key="GridCell.Tab.0.bottom" timestamp="1586198349636">
+    <state width="1398" height="316" key="GridCell.Tab.0.bottom" timestamp="1586278932515">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
-    <state width="1398" height="451" key="GridCell.Tab.0.bottom/0.23.1440.877@0.23.1440.877" timestamp="1586198349636" />
-    <state width="1398" height="451" key="GridCell.Tab.0.center" timestamp="1586198349634">
+    <state width="1398" height="316" key="GridCell.Tab.0.bottom/0.23.1440.877@0.23.1440.877" timestamp="1586278932515" />
+    <state width="1398" height="316" key="GridCell.Tab.0.center" timestamp="1586278932515">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
-    <state width="1398" height="451" key="GridCell.Tab.0.center/0.23.1440.877@0.23.1440.877" timestamp="1586198349634" />
-    <state width="1398" height="451" key="GridCell.Tab.0.left" timestamp="1586198349634">
+    <state width="1398" height="316" key="GridCell.Tab.0.center/0.23.1440.877@0.23.1440.877" timestamp="1586278932515" />
+    <state width="1398" height="316" key="GridCell.Tab.0.left" timestamp="1586278932514">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
-    <state width="1398" height="451" key="GridCell.Tab.0.left/0.23.1440.877@0.23.1440.877" timestamp="1586198349634" />
-    <state width="1398" height="451" key="GridCell.Tab.0.right" timestamp="1586198349635">
+    <state width="1398" height="316" key="GridCell.Tab.0.left/0.23.1440.877@0.23.1440.877" timestamp="1586278932514" />
+    <state width="1398" height="316" key="GridCell.Tab.0.right" timestamp="1586278932515">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
-    <state width="1398" height="451" key="GridCell.Tab.0.right/0.23.1440.877@0.23.1440.877" timestamp="1586198349635" />
+    <state width="1398" height="316" key="GridCell.Tab.0.right/0.23.1440.877@0.23.1440.877" timestamp="1586278932515" />
     <state x="424" y="327" key="com.intellij.jarRepository.RepositoryAttachDialog-DOWNLOAD" timestamp="1585929850650">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
     <state x="424" y="327" key="com.intellij.jarRepository.RepositoryAttachDialog-DOWNLOAD/0.23.1440.877@0.23.1440.877" timestamp="1585929850650" />
-    <state x="499" y="160" key="com.intellij.testIntegration.createTest.CreateTestDialog" timestamp="1585854135333">
+    <state x="499" y="160" key="com.intellij.testIntegration.createTest.CreateTestDialog" timestamp="1586201835847">
       <screen x="0" y="23" width="1440" height="877" />
     </state>
-    <state x="499" y="160" key="com.intellij.testIntegration.createTest.CreateTestDialog/0.23.1440.877@0.23.1440.877" timestamp="1585854135333" />
+    <state x="499" y="160" key="com.intellij.testIntegration.createTest.CreateTestDialog/0.23.1440.877@0.23.1440.877" timestamp="1586201835847" />
   </component>
 </project>
