+++ src/BinaryTree/Tree.java
@@ -1,0 +1,42 @@
+package BinaryTree;
+
+public class Tree {
+    public Tree(Node node) {
+    }
+
+    static class Node {
+        int value;
+        Node left, right;
+
+        Node(int value){
+            this.value = value;
+            left = null;
+            right = null;
+        }
+    }
+
+    public void insert(Node node, int value) {
+        if (value < node.value) {
+            if (node.left != null) {
+                insert(node.left, value);
+            } else {
+                System.out.println(" Inserted " + value + " to left of " + node.value);
+                node.left = new Node(value);
+            }
+        } else if (value > node.value) {
+            if (node.right != null) {
+                insert(node.right, value);
+            } else {
+                System.out.println("  Inserted " + value + " to right of " + node.value);
+                node.right = new Node(value);
+            }
+        }
+    }
+    public void traverseInOrder(Node node) {
+        if (node != null) {
+            traverseInOrder(node.left);
+            System.out.print(" " + node.value);
+            traverseInOrder(node.right);
+        }
+    }
+}
